release: 0.0.1a19.dev20250619
* Update publish_to_pypi.yml


* Update publish_to_pypi.yml


* build: Update test scrtip


* build: Update to bootstrap@5.3.6


* fix: Fix default theme colors


* fix: Fix incorrect route


* feat: Include white theme


* refactor: Themes


* refactor: Config


* chore: Include prettier


* fix: Remove unused import


* refactor: Update HTML templates


* build: Update dev scripts


* refactor: Prettier


* fix:


* fix:


* fix: violates foreign key constraint


* fix:


* fix:


* fix


* fix:


* fix:


* fix:


* fix:


* build: Ignore private templates


* refactor: cli interface


* fix: NameError: name 'with_examples' is not defined


* fix: Name db_migrate already defined on line 448


* fix: Use minnified bootstrap


* refactor: Images


* fix: Update build


* chore: Update build files


* chore: Update build files


* release: 0.0.1a19.dev20250619


---------

+++ now_lms/__init__.py
@@ -108,8 +108,9 @@ from now_lms.misc import (
     concatenar_parametros_a_url,
     markdown_to_clean_hmtl,
 )
-from now_lms.themes import load_theme_variables
+from now_lms.themes import current_theme
 from now_lms.version import VERSION
+from now_lms.vistas._helpers import get_current_course_logo, get_site_logo
 from now_lms.vistas.categories import category
 from now_lms.vistas.certificates import certificate
 from now_lms.vistas.courses import course
@@ -336,7 +337,9 @@ lms_app.jinja_env.globals["adsense_meta"] = get_addsense_meta
 lms_app.jinja_env.globals["adsense_code"] = get_addsense_code
 lms_app.jinja_env.globals["paypal_enabled"] = check_paypal_enabled
 lms_app.jinja_env.globals["paypal_id"] = get_paypal_id
-load_theme_variables(lms_app)
+lms_app.jinja_env.globals["current_theme"] = current_theme
+lms_app.jinja_env.globals["course_logo"] = get_current_course_logo
+lms_app.jinja_env.globals["site_logo"] = get_site_logo
 
 
 # ---------------------------------------------------------------------------------------
@@ -418,12 +421,12 @@ def command(as_module=False) -> None:  # pragma: no cover
 @lms_app.cli.command()
 @click.option("--with-examples", is_flag=True, default=False, help="Load example data at setup.")
 @click.option("--with-testdata", is_flag=True, default=False, help="Load data for testing.")
-def setup(with_examples=False, with_testdata=False):  # pragma: no cover
-    """Inicia al aplicacion."""
+def db_init(with_examples=False, with_testdata=False):  # pragma: no cover
+    """Init a new database."""
     with lms_app.app_context():
         from now_lms.db.tools import database_is_populated
 
-        if database_is_populated(lms_app):
+        if not database_is_populated(lms_app):
 
             initial_setup(with_examples)
             if with_testdata:
@@ -435,39 +438,63 @@ def setup(with_examples=False, with_testdata=False):  # pragma: no cover
 
 
 @lms_app.cli.command()
-def release():  # pragma: no cover
-    """Devuelve la versión actual del programa."""
-    print(VERSION)
+def db_seed():
+    """Setup a new develoment database."""
+
+    from now_lms.db.data_test import crear_data_para_pruebas
+
+    with lms_app.app_context():
+        initial_setup(with_examples=True)
+        crear_data_para_pruebas()
 
 
 @lms_app.cli.command()
-def upgrade_db():  # pragma: no cover
-    """Actualiza esquema de base de datos."""
+def db_backup():  # pragma: no cover
+    """Make a backup of system data."""
+    pass
+
+
+@lms_app.cli.command()
+def db_backup_restore():  # pragma: no cover
+    """Restore the system from a backup."""
+    pass
+
+
+@lms_app.cli.command()
+def db_migrate():  # pragma: no cover
+    """Update dabatase schema."""
     alembic.upgrade()
 
 
 @lms_app.cli.command()
-def deletedb():  # pragma: no cover
-    """Elimina base de datos."""
-    if DESARROLLO:
-        with lms_app.app_context():
+def db_drop():  # pragma: no cover
+    """Delete database schema and all the data in it."""
+    with lms_app.app_context():
+        if click.confirm("This will delete the database and all the data on it. Do you want to continue?", abort=True):
             database.drop_all()
 
 
 @lms_app.cli.command()
 @click.option("--with-examples", is_flag=True, default=False, help="Load example data at setup.")
 @click.option("--with-tests", is_flag=True, default=False, help="Load data for testing.")
-def resetdb(with_examples=False, with_tests=False) -> None:  # pragma: no cover
-    """Elimina la base de datos actual e inicia una nueva."""
+def db_reset(with_examples=False, with_tests=False) -> None:  # pragma: no cover
+    """Drop the system database and populate with init a new one."""
     with lms_app.app_context():
-        cache.clear()
-        database.drop_all()
-        initial_setup(with_examples, with_tests)
+        if click.confirm("This will delete the current database and all the data on it. Do you want to continue?", abort=True):
+            cache.clear()
+            database.drop_all()
+            initial_setup(with_examples, with_tests)
+
+
+@lms_app.cli.command()
+def version():  # pragma: no cover
+    """Return the current versión of the software."""
+    click.echo(VERSION)
 
 
 @lms_app.cli.command()
 def serve():  # pragma: no cover
-    """Servidor WSGI predeterminado."""
+    """Serve NOW LMS with the default WSGi server."""
     from waitress import serve as server
 
     if environ.get("LMS_PORT"):
